+++ src/kard/commons/types/progress_bar_segment.py
@@ -3,7 +3,7 @@
 import typing
 
 import pydantic
-from ....core.pydantic_utilities import IS_PYDANTIC_V2, UniversalBaseModel
+from ...core.pydantic_utilities import IS_PYDANTIC_V2, UniversalBaseModel
 from .progress_bar_segment_label import ProgressBarSegmentLabel
 from .progress_bar_segment_position import ProgressBarSegmentPosition
 from .progress_bar_segment_selection import ProgressBarSegmentSelection