+++ src/kard/users/rewards/types/__init__.py
@@ -11,7 +11,6 @@ if typing.TYPE_CHECKING:
     from .asset import Asset
     from .batches_meta import BatchesMeta
     from .batches_response_object import BatchesResponseObject
-    from .button_style import ButtonStyle
     from .category_data import CategoryData
     from .category_fields import CategoryFields
     from .category_identifier import CategoryIdentifier
@@ -21,8 +20,6 @@ if typing.TYPE_CHECKING:
     from .commission import Commission
     from .component_type import ComponentType
     from .coordinates import Coordinates
-    from .cta_action import CtaAction
-    from .cta_component import CtaComponent
     from .eligibility_location_address import EligibilityLocationAddress
     from .eligibility_location_included import EligibilityLocationIncluded
     from .eligibility_offer_included import EligibilityOfferIncluded
@@ -34,12 +31,7 @@ if typing.TYPE_CHECKING:
     from .location_relationships import LocationRelationships
     from .location_sort_options import LocationSortOptions
     from .locations_response_object import LocationsResponseObject
-    from .logo_flare import LogoFlare
-    from .logo_flare_badge import LogoFlareBadge
-    from .logo_flare_badge_position import LogoFlareBadgePosition
-    from .logo_flare_border_color import LogoFlareBorderColor
     from .offer_common_fields import OfferCommonFields
-    from .offer_components import OfferComponents
     from .offer_data_union import OfferDataUnion, OfferDataUnion_StandardOffer
     from .offer_relationship import OfferRelationship
     from .offer_sort_options import OfferSortOptions
@@ -51,16 +43,6 @@ if typing.TYPE_CHECKING:
     from .placement_batch_attributes import PlacementBatchAttributes
     from .placement_batch_data import PlacementBatchData
     from .placement_content_response import PlacementContentResponse
-    from .progress_bar import ProgressBar
-    from .progress_bar_label_pair import ProgressBarLabelPair
-    from .progress_bar_labels import ProgressBarLabels
-    from .progress_bar_segment import ProgressBarSegment
-    from .progress_bar_segment_label import ProgressBarSegmentLabel
-    from .progress_bar_segment_position import ProgressBarSegmentPosition
-    from .progress_bar_segment_progress import ProgressBarSegmentProgress
-    from .progress_bar_segment_selection import ProgressBarSegmentSelection
-    from .progress_bar_segment_separator import ProgressBarSegmentSeparator
-    from .progress_bar_segments import ProgressBarSegments
     from .standard_offer import StandardOffer
     from .standard_offer_core import StandardOfferCore
     from .standard_offer_fields import StandardOfferFields
@@ -70,7 +52,6 @@ _dynamic_imports: typing.Dict[str, str] = {
     "Asset": ".asset",
     "BatchesMeta": ".batches_meta",
     "BatchesResponseObject": ".batches_response_object",
-    "ButtonStyle": ".button_style",
     "CategoryData": ".category_data",
     "CategoryFields": ".category_fields",
     "CategoryIdentifier": ".category_identifier",
@@ -80,8 +61,6 @@ _dynamic_imports: typing.Dict[str, str] = {
     "Commission": ".commission",
     "ComponentType": ".component_type",
     "Coordinates": ".coordinates",
-    "CtaAction": ".cta_action",
-    "CtaComponent": ".cta_component",
     "EligibilityLocationAddress": ".eligibility_location_address",
     "EligibilityLocationIncluded": ".eligibility_location_included",
     "EligibilityOfferIncluded": ".eligibility_offer_included",
@@ -93,12 +72,7 @@ _dynamic_imports: typing.Dict[str, str] = {
     "LocationRelationships": ".location_relationships",
     "LocationSortOptions": ".location_sort_options",
     "LocationsResponseObject": ".locations_response_object",
-    "LogoFlare": ".logo_flare",
-    "LogoFlareBadge": ".logo_flare_badge",
-    "LogoFlareBadgePosition": ".logo_flare_badge_position",
-    "LogoFlareBorderColor": ".logo_flare_border_color",
     "OfferCommonFields": ".offer_common_fields",
-    "OfferComponents": ".offer_components",
     "OfferDataUnion": ".offer_data_union",
     "OfferDataUnion_StandardOffer": ".offer_data_union",
     "OfferRelationship": ".offer_relationship",
@@ -111,16 +85,6 @@ _dynamic_imports: typing.Dict[str, str] = {
     "PlacementBatchAttributes": ".placement_batch_attributes",
     "PlacementBatchData": ".placement_batch_data",
     "PlacementContentResponse": ".placement_content_response",
-    "ProgressBar": ".progress_bar",
-    "ProgressBarLabelPair": ".progress_bar_label_pair",
-    "ProgressBarLabels": ".progress_bar_labels",
-    "ProgressBarSegment": ".progress_bar_segment",
-    "ProgressBarSegmentLabel": ".progress_bar_segment_label",
-    "ProgressBarSegmentPosition": ".progress_bar_segment_position",
-    "ProgressBarSegmentProgress": ".progress_bar_segment_progress",
-    "ProgressBarSegmentSelection": ".progress_bar_segment_selection",
-    "ProgressBarSegmentSeparator": ".progress_bar_segment_separator",
-    "ProgressBarSegments": ".progress_bar_segments",
     "StandardOffer": ".standard_offer",
     "StandardOfferCore": ".standard_offer_core",
     "StandardOfferFields": ".standard_offer_fields",
@@ -154,7 +118,6 @@ __all__ = [
     "Asset",
     "BatchesMeta",
     "BatchesResponseObject",
-    "ButtonStyle",
     "CategoryData",
     "CategoryFields",
     "CategoryIdentifier",
@@ -164,8 +127,6 @@ __all__ = [
     "Commission",
     "ComponentType",
     "Coordinates",
-    "CtaAction",
-    "CtaComponent",
     "EligibilityLocationAddress",
     "EligibilityLocationIncluded",
     "EligibilityOfferIncluded",
@@ -177,12 +138,7 @@ __all__ = [
     "LocationRelationships",
     "LocationSortOptions",
     "LocationsResponseObject",
-    "LogoFlare",
-    "LogoFlareBadge",
-    "LogoFlareBadgePosition",
-    "LogoFlareBorderColor",
     "OfferCommonFields",
-    "OfferComponents",
     "OfferDataUnion",
     "OfferDataUnion_StandardOffer",
     "OfferRelationship",
@@ -195,16 +151,6 @@ __all__ = [
     "PlacementBatchAttributes",
     "PlacementBatchData",
     "PlacementContentResponse",
-    "ProgressBar",
-    "ProgressBarLabelPair",
-    "ProgressBarLabels",
-    "ProgressBarSegment",
-    "ProgressBarSegmentLabel",
-    "ProgressBarSegmentPosition",
-    "ProgressBarSegmentProgress",
-    "ProgressBarSegmentSelection",
-    "ProgressBarSegmentSeparator",
-    "ProgressBarSegments",
     "StandardOffer",
     "StandardOfferCore",
     "StandardOfferFields",